+++ wsuwp-a11y-status.php
@@ -1,7 +1,7 @@
 <?php
 /*
 Plugin Name: WSUWP A11y Status
-Version: 0.7.0
+Version: 0.8.0
 Description: A plugin to view users' WSU Accessibility Training status in the Admin area.
 Author: washingtonstateuniversity, Adam Turner
 Author URI: https://github.com/washingtonstateuniversity/
@@ -37,11 +37,12 @@ register_uninstall_hook( __FILE__, array( 'WSUWP_A11y_Status', 'uninstall' ) );
  *
  * @since 0.1.0
  *
- * @return object An instance of WSUWP_A11y_Status
+ * @return WSUWP_A11y_Status An instance of the WSUWP_A11y_Status class.
  */
 function load_wsuwp_a11y_status() {
 	$wsuwp_a11y_status = WSUWP_A11y_Status::get_instance();
 	$wsuwp_a11y_status->setup_hooks();
+	$wsuwp_a11y_status->set_properties( __FILE__ );
 
 	return $wsuwp_a11y_status;
 }